Only save configuration values defined to be persisted
The entire application config can be persisted to disk but not every
item is meant to be saved to disk as they are designed to be
options the user passes during runtime and not enabled every
single run. This ensures only configuration defined in the
DEFAULTS hash is persisted to disk.

This also prevents the false assumption that changing a saved value
in the config will affect the next run which is not always true for
every value.

(cherry picked from commit f30226f5ecbac4e05bef45a8c5b5c0c1ca2fd4e4)